Move DayOfWeek locale-related functions
