+++ core/src/commonMain/kotlin/io/islandtime/locale/Locale.kt
@@ -1,7 +1,5 @@
 package io.islandtime.locale
 
-import io.islandtime.DayOfWeek
-
 /**
  * A locale.
  *
